[refactor] Rest Docs 리팩토링 및 release 0.0.1(#31)

+++ src/main/java/upbrella/be/rent/dto/request/ReturnUmbrellaByUserRequest.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ReturnUmbrellaByUserRequest {
 
-    private int umbrellaId;
-    private int storeId;
+    private long umbrellaId;
+    private long storeId;
     private String improvement;
 }
